BTEUK-158: Expand gui functionality to allow for temporary guis.
Add a contructor to the LocationMenu to create a temporary version for nearby locations.

+++ src/main/java/me/bteuk/network/gui/navigation/LocationMenu.java
@@ -27,131 +27,44 @@ import static me.bteuk.network.utils.NetworkConfig.CONFIG;
 public class LocationMenu extends Gui {
 
     private LinkedHashSet<String> locations;
-
-    private int page;
-
-    private final String type;
-    private final String returnMenu;
-    private final NetworkUser u;
-
-    //Method to determine the search parameters when getting the locations to display in the menu.
-    private LinkedHashSet<String> getLocations() {
-
-        switch (type) {
-
-            //UK Categories (Excluding England)
-            case "Scotland" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='SCOTLAND' ORDER BY location ASC;"));}
-            case "Wales" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='WALES' ORDER BY location ASC;"));}
-            case "Northern Ireland" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='NORTHERN_IRELAND' ORDER BY location ASC;"));}
-            case "Other" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='OTHER' ORDER BY location ASC;"));}
-            case "Suggested" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE suggested=1 ORDER BY location ASC;"));}
-
-            //Regions of England
-            case "Yorkshire" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='YORKSHIRE' ORDER BY location ASC;"));}
-            case "West Midlands" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='WEST_MIDLANDS' ORDER BY location ASC;"));}
-            case "London" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='LONDON' ORDER BY location ASC;"));}
-            case "East Midlands" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='EAST_MIDLANDS' ORDER BY location ASC;"));}
-            case "East of England" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='EAST_OF_ENGLAND' ORDER BY location ASC;"));}
-            case "North West" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='NORTH_WEST' ORDER BY location ASC;"));}
-            case "North East" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='NORTH_EAST' ORDER BY location ASC;"));}
-            case "South East" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='SOUTH_EAST' ORDER BY location ASC;"));}
-            case "South West" -> {return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='SOUTH_WEST' ORDER BY location ASC;"));}
-
-            //Nearby locations to the player.
-            case "Nearby" -> {return getNearbyLocations();}
-
-            //This is for the 'search' case where the location is based on the user input.
-            default -> {return searchLocations();}
-
-        }
-    }
-
-    private LinkedHashSet<String> getNearbyLocations() {
-
-        Location l = u.player.getLocation();
-        String worldName = l.getWorld().getName();
-
-        //Check if world is in plotsystem.
-        if (Network.getInstance().getPlotSQL().hasRow("SELECT name FROM location_data WHERE name='" + worldName + "';")) {
-
-            //Add coordinate transformation.
-            l = new Location(
-                    Bukkit.getWorld(worldName),
-                    u.player.getLocation().getX() - Network.getInstance().getPlotSQL().getInt("SELECT xTransform FROM location_data WHERE name='" + worldName + "';"),
-                    u.player.getLocation().getY(),
-                    u.player.getLocation().getZ() - Network.getInstance().getPlotSQL().getInt("SELECT zTransform FROM location_data WHERE name='" + worldName + "';"),
-                    u.player.getLocation().getYaw(),
-                    u.player.getLocation().getPitch()
-            );
-
-        }
-
-        return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location_data.location FROM location_data INNER JOIN coordinates ON location_data.coordinate=coordinates.id " +
-                "WHERE ((((coordinates.x/1000)-" + (l.getX() / 1000) + ")*((coordinates.x/1000)-" + (l.getX() / 1000) + ")) + " +
-                "(((coordinates.z/1000)-" + (l.getZ() / 1000) + ")*((coordinates.z/1000)-" + (l.getZ() / 1000) + "))) < " +
-                (CONFIG.getInt("navigation_radius") * CONFIG.getInt("navigation_radius")) +
-                " ORDER BY ((((coordinates.x/1000)-" + (l.getX() / 1000) + ")*((coordinates.x/1000)-" + (l.getX() / 1000) + ")) + " +
-                "(((coordinates.z/1000)-" + (l.getZ() / 1000) + ")*((coordinates.z/1000)-" + (l.getZ() / 1000) + "))) ASC;"));
-
-    }
-
-    private LinkedHashSet<String> searchLocations() {
-
-        //Search for locations that include this phrase.
-        ArrayList<String> locations = Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE location LIKE '%" + type + "%';");
-
-        //Also search for any regions or counties.
-        for (Counties county : Counties.values()) {
-            if (StringUtils.containsIgnoreCase(county.label, type)) {
-
-                locations.addAll(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE subcategory='" + county.region + "';"));
-
-            }
-        }
-
-        for (Categories category : Categories.values()) {
-            if (StringUtils.containsIgnoreCase(category.label, type)) {
-
-                locations.addAll(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='" + category + "';"));
-
-            }
-        }
-
-        locations.sort(Comparator.naturalOrder());
-
-        return new LinkedHashSet<>(locations);
-
-    }
-
-    //Function to get the gui for the return button.
-    private Gui getReturnGui() {
-        if (returnMenu.equals("England")) {
-
-            return new EnglandMenu();
-
-        } else {
-
-            return new ExploreGui(u);
-
-        }
-    }
+    private int page = 1;
+    private String type = "Nearby";
+    private String returnMenu = null;
+    private Location l = null;
 
     public LocationMenu(String title, NetworkUser u, String type, String returnMenu) {
-
         super(45, Component.text(title, NamedTextColor.AQUA, TextDecoration.BOLD));
-
-        this.u = u;
 
         this.type = type;
         this.returnMenu = returnMenu;
 
+        // If the type is Nearby.
+        if (type.equals("Nearby")) {
+            l = u.getLocationWithCoordinateTransform();
+        }
+
+        createLocationMenu();
+    }
+
+    /**
+     * Create a temporary gui that will delete itself on close.
+     * The gui will show all nearby locations to the inputted location.
+     */
+    public LocationMenu(String title, Location l) {
+        super(45, Component.text(title, NamedTextColor.AQUA, TextDecoration.BOLD));
+
+        setDeleteOnClose(true);
+        this.l = l;
+
+        createLocationMenu();
+    }
+
+    /**
+     * Generic location menu method that shares functions from other constructors.
+     */
+    private void createLocationMenu() {
         this.locations = getLocations();
-
-        //On initialization the page is always 1.
-        page = 1;
-
         createGui();
-
     }
 
     private void createGui() {
@@ -279,22 +192,24 @@ public class LocationMenu extends Gui {
 
         }
 
-        //Return
-        setItem(44, Utils.createItem(Material.SPRUCE_DOOR, 1,
-                        Utils.title("Return"),
-                        Utils.line("Open the exploration menu.")),
-                u ->
+        //Return (optional)
+        if (returnMenu != null) {
+            setItem(44, Utils.createItem(Material.SPRUCE_DOOR, 1,
+                            Utils.title("Return"),
+                            Utils.line("Open the exploration menu.")),
+                    u ->
 
-                {
+                    {
 
-                    //Delete this gui.
-                    this.delete();
+                        //Delete this gui.
+                        this.delete();
 
-                    //Switch to navigation menu.
-                    u.mainGui = getReturnGui();
-                    u.mainGui.open(u);
+                        //Switch to navigation menu.
+                        u.mainGui = getReturnGui(u);
+                        u.mainGui.open(u);
 
-                });
+                    });
+        }
     }
 
     public void refresh() {
@@ -306,8 +221,8 @@ public class LocationMenu extends Gui {
 
         //Check if page has content.
         //Else set it to the maximum possible.
-        if (Math.ceil(locations.size()/21.0) < page) {
-            page = locations.size()/21;
+        if (Math.ceil(locations.size() / 21.0) < page) {
+            page = locations.size() / 21;
         }
 
         createGui();
@@ -316,5 +231,119 @@ public class LocationMenu extends Gui {
 
     public boolean isEmpty() {
         return locations.isEmpty();
+    }
+
+    //Method to determine the search parameters when getting the locations to display in the menu.
+    private LinkedHashSet<String> getLocations() {
+
+        switch (type) {
+
+            //UK Categories (Excluding England)
+            case "Scotland" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='SCOTLAND' ORDER BY location ASC;"));
+            }
+            case "Wales" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='WALES' ORDER BY location ASC;"));
+            }
+            case "Northern Ireland" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='NORTHERN_IRELAND' ORDER BY location ASC;"));
+            }
+            case "Other" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='OTHER' ORDER BY location ASC;"));
+            }
+            case "Suggested" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE suggested=1 ORDER BY location ASC;"));
+            }
+
+            //Regions of England
+            case "Yorkshire" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='YORKSHIRE' ORDER BY location ASC;"));
+            }
+            case "West Midlands" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='WEST_MIDLANDS' ORDER BY location ASC;"));
+            }
+            case "London" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='LONDON' ORDER BY location ASC;"));
+            }
+            case "East Midlands" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='EAST_MIDLANDS' ORDER BY location ASC;"));
+            }
+            case "East of England" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='EAST_OF_ENGLAND' ORDER BY location ASC;"));
+            }
+            case "North West" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='NORTH_WEST' ORDER BY location ASC;"));
+            }
+            case "North East" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='NORTH_EAST' ORDER BY location ASC;"));
+            }
+            case "South East" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='SOUTH_EAST' ORDER BY location ASC;"));
+            }
+            case "South West" -> {
+                return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='ENGLAND' AND subcategory='SOUTH_WEST' ORDER BY location ASC;"));
+            }
+
+            //Nearby locations to the player.
+            case "Nearby" -> {
+                return getNearbyLocations();
+            }
+
+            //This is for the 'search' case where the location is based on the user input.
+            default -> {
+                return searchLocations();
+            }
+
+        }
+    }
+
+    private LinkedHashSet<String> getNearbyLocations() {
+        return new LinkedHashSet<>(Network.getInstance().globalSQL.getStringList("SELECT location_data.location FROM location_data INNER JOIN coordinates ON location_data.coordinate=coordinates.id " +
+                "WHERE ((((coordinates.x/1000)-" + (l.getX() / 1000) + ")*((coordinates.x/1000)-" + (l.getX() / 1000) + ")) + " +
+                "(((coordinates.z/1000)-" + (l.getZ() / 1000) + ")*((coordinates.z/1000)-" + (l.getZ() / 1000) + "))) < " +
+                (CONFIG.getInt("navigation_radius") * CONFIG.getInt("navigation_radius")) +
+                " ORDER BY ((((coordinates.x/1000)-" + (l.getX() / 1000) + ")*((coordinates.x/1000)-" + (l.getX() / 1000) + ")) + " +
+                "(((coordinates.z/1000)-" + (l.getZ() / 1000) + ")*((coordinates.z/1000)-" + (l.getZ() / 1000) + "))) ASC;"));
+    }
+
+    private LinkedHashSet<String> searchLocations() {
+
+        //Search for locations that include this phrase.
+        ArrayList<String> locations = Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE location LIKE '%" + type + "%';");
+
+        //Also search for any regions or counties.
+        for (Counties county : Counties.values()) {
+            if (StringUtils.containsIgnoreCase(county.label, type)) {
+
+                locations.addAll(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE subcategory='" + county.region + "';"));
+
+            }
+        }
+
+        for (Categories category : Categories.values()) {
+            if (StringUtils.containsIgnoreCase(category.label, type)) {
+
+                locations.addAll(Network.getInstance().globalSQL.getStringList("SELECT location FROM location_data WHERE category='" + category + "';"));
+
+            }
+        }
+
+        locations.sort(Comparator.naturalOrder());
+
+        return new LinkedHashSet<>(locations);
+
+    }
+
+    //Function to get the gui for the return button.
+    private Gui getReturnGui(NetworkUser u) {
+        if (returnMenu.equals("England")) {
+
+            return new EnglandMenu();
+
+        } else {
+
+            return new ExploreGui(u);
+
+        }
     }
 }
